添加自定义drawable

+++ convenientui/src/main/java/com/scorpio/ui/util/DensityUtil.java
@@ -16,9 +16,9 @@ public class DensityUtil {
      * @param dpValue
      * @return
      */
-    public static int dip2px(Context context, float dpValue) {
+    public static float dip2px(Context context, float dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (dpValue * scale + 0.5f);
+        return dpValue * scale + 0.5f;
     }
 
     /**
@@ -28,9 +28,9 @@ public class DensityUtil {
      * @param pxValue
      * @return
      */
-    public static int px2dip(Context context, float pxValue) {
+    public static float px2dip(Context context, float pxValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (pxValue / scale + 0.5f);
+        return pxValue / scale + 0.5f;
     }
 
     /**
@@ -39,14 +39,14 @@ public class DensityUtil {
      * @param spValue
      * @return
      */
-    public static int sp2px(Context context, float spValue) {
+    public static float sp2px(Context context, float spValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (spValue * scale + 0.5f);
+        return spValue * scale + 0.5f;
     }
 
     // 将px值转换为sp值，保证文字大小不变
-    public static int px2sp(Context context, float pxValue) {
+    public static float px2sp(Context context, float pxValue) {
         final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (pxValue / fontScale + 0.5f);
+        return pxValue / fontScale + 0.5f;
     }
 }
